+++ src/Cooked/MockChain/Balancing.hs
@@ -236,8 +236,18 @@ setFeeAndBalance balanceWallet skel0 = do
     if txOptEnsureMinAda . txSkelOpts $ skel0
       then ensureTxSkelOutsMinAda skel0
       else return skel0
-  -- all UTxOs belonging to the balancing public key
-  balancePKUtxos <- Map.fromList <$> utxosAtLedger (walletAddress balanceWallet)
+
+  -- Candidate UTxOs to spend for balancing
+  balancePKUtxos <-
+    Map.fromList
+      <$> let balanceWalletAddress = walletAddress balanceWallet
+              noDatumPredicate = (\case PV2.NoOutputDatum -> True; _ -> False) . outputOutputDatum . txOutV2FromLedger
+           in case txOptBalancingUtxos . txSkelOpts $ skel0 of
+                BalancingUtxosAll -> utxosAtLedger balanceWalletAddress
+                BalancingUtxosDatumless -> runUtxoSearch (utxosAtLedgerSearch balanceWalletAddress `filterWithPred` noDatumPredicate)
+                BalancingUtxosWith txOutRefs -> filter ((`elem` txOutRefs) . fst) <$> utxosAtLedger balanceWalletAddress
+                BalancingUtxosWithout txOutRefs -> filter (not . (`elem` txOutRefs) . fst) <$> utxosAtLedger balanceWalletAddress
+
   -- all UTxOs that the txSkel consumes.
   txSkelUtxos <- txSkelInputUtxos skel
   -- all UTxOs that the txSkel references.