+++ src/Cooked/Skeleton.hs
@@ -20,6 +20,7 @@ module Cooked.Skeleton
     BalancingWallet (..),
     RawModTx (..),
     EmulatorParamsModification (..),
+    BalancingUtxos (..),
     applyEmulatorParamsModification,
     applyRawModOnBalancedTx,
     TxOpts (..),
@@ -101,6 +102,7 @@ import qualified Plutus.Script.Utils.Value as Pl hiding (adaSymbol, adaToken)
 import qualified Plutus.V1.Ledger.Interval as Pl
 import qualified Plutus.V2.Ledger.Api as Pl hiding (TxOut, adaSymbol, adaToken)
 import qualified Plutus.V2.Ledger.Tx as Pl
+import qualified Plutus.V2.Ledger.Tx as Pl2
 import qualified PlutusTx.Prelude as Pl
 import Test.QuickCheck (NonZero (..))
 import Type.Reflection
@@ -193,6 +195,21 @@ applyEmulatorParamsModification :: Maybe EmulatorParamsModification -> Emulator.
 applyEmulatorParamsModification (Just (EmulatorParamsModification f)) = f
 applyEmulatorParamsModification Nothing = id
 
+-- | Describes which UTxOs of the balancing wallet can be spent for balancing.
+data BalancingUtxos
+  = -- | Use all UTxOs (default)
+    BalancingUtxosAll
+  | -- | Use all UTxOs without datum
+    BalancingUtxosDatumless
+  | -- | Use only the provided UTxOs
+    BalancingUtxosWith [Pl2.TxOutRef]
+  | -- | Do not use the provided UTxOs
+    BalancingUtxosWithout [Pl2.TxOutRef]
+  deriving (Eq, Ord, Show)
+
+instance Default BalancingUtxos where
+  def = BalancingUtxosAll
+
 -- | Set of options to modify the behavior of generating and validating some transaction.
 data TxOpts = TxOpts
   { -- | Performs an adjustment to unbalanced transactions, making sure every
@@ -249,6 +266,11 @@ data TxOpts = TxOpts
     --
     -- Default is 'BalanceWithFirstSigner'.
     txOptBalanceWallet :: BalancingWallet,
+    -- | Describes which UTxOs of the balancing wallet can be spent for
+    -- balancing. This is useful to put aside some UTxOs you want to spend in a
+    -- specific context later and prevent premature spending during balancing
+    -- of previous transactions.
+    txOptBalancingUtxos :: BalancingUtxos,
     -- | Apply an arbitrary modification to the protocol parameters that are
     -- used to balance and submit the transaction. This is
     -- obviously a very unsafe thing to do if you want to preserve
@@ -276,6 +298,7 @@ instance Default TxOpts where
         txOptBalance = True,
         txOptBalanceOutputPolicy = def,
         txOptBalanceWallet = def,
+        txOptBalancingUtxos = def,
         txOptEmulatorParamsModification = Nothing
       }